cosmos-db-migration-utility/src/migrator-app/requirements.txt
pymongo==3.10.1
PyYAML==5.3
boto3==1.12.17
argparse==1.4.0